requirements.txt
aenum==2.2.6
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cssselect2==0.5.0
defusedxml==0.7.1
geojson==2.5.0
httplib2==0.20.4
idna==3.3
numpy==1.22.3
opencv-python==4.5.5.64
paho-mqtt==1.6.1
Pillow==9.1.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pyowm==3.3.0
pyparsing==3.0.8
PySocks==1.7.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rsa==4.8
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
Telethon==1.24.0
tinycss2==1.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
webencodings==0.5.1